+++ src/pl/plajer/buildbattle3/handlers/language/LanguageMigrator.java
@@ -43,7 +43,7 @@ import pl.plajer.buildbattle3.utils.MessageUtils;
  */
 public class LanguageMigrator {
 
-  public static final int LANGUAGE_FILE_VERSION = 4;
+  public static final int LANGUAGE_FILE_VERSION = 5;
   public static final int CONFIG_FILE_VERSION = 2;
 
   private static Main plugin = JavaPlugin.getPlugin(Main.class);